[REFACTOR] : 코드 분리
