+++ src/pages/MeetingRoom/views/MeetingRoom.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { contentsCenter } from './MapStyle';
-// import MapNav from '../../../components/MapNav/views/MapNav';
+// import MapNav from '../../../components/MapNav/MapNav';
 import SideBar from '../../../components/SideBar/views/SideBar';
 import C3Map from './C3Map';
