+++ src/styles/arrow.tsx
@@ -1,0 +1,8 @@
+import { css } from '@emotion/react'
+import tw from "twin.macro";
+
+const arrow = css`
+    ${tw`mx-3 w-12 h-12 bg-cover`}
+`
+
+export default arrow;
